+++ src/main/java/com/hendisantika/adminlte/repository/CustomersRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomersRepository extends JpaRepository<Customers, Long> {
+public interface CustomersRepository extends JpaRepository<Customers, Integer> {
 
 }
